+++ src/test/java/lotto/data/LottoTest.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.data;
 
 import lotto.constants.ErrorCode;
 import lotto.data.Lotto;
